fix fetch order
